Modified reminders logic to use setTimeout
