+++ package.json
@@ -54,8 +54,7 @@
         "lint-staged": "^8.1.3",
         "mocha": "5.2.0",
         "prettier": "^1.16.4",
-        "prop-types": "^15.6.2",
-        "sw-precache": "^5.2.1"
+        "prop-types": "^15.6.2"
     },
     "browserslist": [
         ">0.2%",
